Flask_Docker_Project/requirements.txt
flask
wtforms
wtforms.validators
mysql.connector
flask_fontawesome 